Refactor mixtape routes and update import paths for better organization
